vinacode/wav2vec2-large-xlsr-53-vietnamese-main/requirements.txt
torchaudio
librosa
transformers
datasets